requirements.txt
numpy==1.26.4
pandas==2.2.2
polars==0.20.26
dash==2.17.0
plotly-geo==1.0.0
geoarrow-rust-core==0.2.0
pyogrio==0.8.0
geopandas==0.14.4
pyarrow==16.1.0
gunicorn==22.0.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlalchemy==2.0.31
geoalchemy2==0.15.1